+++ packages/jsdestroy/src/defineDestroyProperties.js
@@ -6,7 +6,7 @@ function returnFalse() {
 
 /**
  * 为对象或原型链定义destroy相关Property
- * @param {Object||Object.prototype} target 对象或原型链
+ * @param {Object|Object.prototype} target 对象或原型链
  * @param {Function} destroyFunc 外部传入destroy方法，可不填
  * @returns {void}
  */
@@ -18,15 +18,29 @@ function defineDestroyProperties(target, destroyFunc) {
             writable: true,
             value: returnFalse,
         },
+        onDestroy: {
+            configurable: true,
+            enumerable: true,
+            writable: true,
+            value: function () {
+                // console.warn('onDestroy must be overwrited by subclass.');
+            }
+        },
         destroy: {
             configurable: true,
             enumerable: true,
             writable: true,
             value: function () {
-                if (destroyFunc)
-                    destroyFunc(this);
-                else
-                    destroyObject(this);
+                if (this.isDestroyed()) {
+                    console.warn('This object was destroyed.', this);
+                } else {
+                    this.onDestroy();
+                    if (destroyFunc)
+                        destroyFunc(this);
+                    else
+                        destroyObject(this);
+                }
+                return this;
             }
         }
     });
